+++ packages/publint/src/index.d.ts
@@ -80,9 +80,8 @@ export type Message =
         expectExtension?: string
       }
     >
-  // TODO: rename EXPORT to EXPORTS (i messed up)
   | BaseMessage<
-      'EXPORT_TYPES_INVALID_FORMAT',
+      'EXPORTS_TYPES_INVALID_FORMAT',
       {
         condition: string
         actualFormat: string
@@ -121,6 +120,7 @@ export type Message =
       }
     >
   | BaseMessage<'LOCAL_DEPENDENCY'>
+  | BaseMessage<'BIN_FILE_NOT_EXECUTABLE'>
 
 export interface PackFile {
   name: string