+++ packages/publint/src/shared/core.js
@@ -28,6 +28,7 @@ import {
   isShorthandRepositoryUrl,
   isShorthandGitHubOrGitLabUrl,
   isDeprecatedGitHubGitUrl,
+  startsWithShebang,
 } from './utils.js'
 
 /**
@@ -444,6 +445,12 @@ export async function core({ pkgDir, vfs, level, strict, _packedFiles }) {
     })
   }
 
+  // check file existence for bin field
+  const [bin, binPkgPath] = getPublishedField(rootPkg, 'bin')
+  if (bin) {
+    crawlBin(bin, binPkgPath)
+  }
+
   await promiseQueue.wait()
 
   if (strict) {
@@ -626,14 +633,16 @@ export async function core({ pkgDir, vfs, level, strict, _packedFiles }) {
 
   /**
    * @param {string} exportsValue
+   * @param {string} exportsKey
+   * @param {Record<string, any>} exports
    */
-  async function getExportsFiles(exportsValue) {
+  async function getExportsFiles(exportsValue, exportsKey, exports) {
     const exportsPath = isRelativePath(exportsValue)
       ? vfs.pathJoin(pkgDir, exportsValue)
       : exportsValue
     const isGlob = exportsValue.includes('*')
     return isGlob
-      ? await exportsGlob(exportsPath, vfs, _packedFiles)
+      ? await exportsGlob(exportsPath, vfs, _packedFiles, exportsKey, exports)
       : [exportsPath]
   }
 
@@ -684,7 +693,12 @@ export async function core({ pkgDir, vfs, level, strict, _packedFiles }) {
         }
 
         const isGlob = exportsValue.includes('*')
-        const exportsFiles = await getExportsFiles(exportsValue)
+        const exportsKey = currentPath[1]
+        const exportsFiles = await getExportsFiles(
+          exportsValue,
+          exportsKey,
+          exports,
+        )
 
         if (isGlob && !exportsFiles.length) {
           messages.push({
@@ -756,7 +770,7 @@ export async function core({ pkgDir, vfs, level, strict, _packedFiles }) {
               return
             }
             // file format checks isn't required for `browser` condition or exports
-            // after the node condtion, as nodejs doesn't use it, only bundlers do,
+            // after the node condition, as nodejs doesn't use it, only bundlers do,
             // which doesn't care of the format
             if (isAfterNodeCondition || currentPath.includes('browser')) return
             const actualFormat = getCodeFormat(fileContent)
@@ -819,9 +833,21 @@ export async function core({ pkgDir, vfs, level, strict, _packedFiles }) {
       if ('types' in exportsValue && exportsKeys[0] !== 'types') {
         // check preceding conditions before the `types` condition, if there are nested
         // conditions, check if they also have the `types` condition. If they do, there's
-        // a good chance those take precendence over this non-first `types` condition, which
-        // is fine and is usually used as fallback instead.
-        const precedingKeys = exportsKeys.slice(0, exportsKeys.indexOf('types'))
+        // a good chance those take precedence over this non-first `types` condition, which
+        // is fine and is usually used as fallback instead. Versioned `types` conditions
+        // are allowed to precede the `types` condition.
+        const precedingKeys = exportsKeys
+          .slice(0, exportsKeys.indexOf('types'))
+          .filter((key) => !key.startsWith('types'))
+
+        // TODO: check that versioned types are valid and the ranges don't overlap.
+
+        // TODO: check the order of the conditions are correct. For example,
+        //  1. Nested conditions
+        //  2. Versioned `types` conditions
+        //  3. `types` condition
+        //  4. All other conditions
+
         let isPrecededByNestedTypesCondition = false
         for (const key of precedingKeys) {
           if (
@@ -832,7 +858,7 @@ export async function core({ pkgDir, vfs, level, strict, _packedFiles }) {
             break
           }
         }
-        if (!isPrecededByNestedTypesCondition) {
+        if (precedingKeys.length > 0 && !isPrecededByNestedTypesCondition) {
           messages.push({
             code: 'EXPORTS_TYPES_SHOULD_BE_FIRST',
             args: {},
@@ -925,42 +951,52 @@ export async function core({ pkgDir, vfs, level, strict, _packedFiles }) {
 
         // keyed strings for seen resolved paths, so we don't trigger duplicate messages for the same thing
         const seenResolvedKeys = new Set()
+        /**
+         * @param {(string | undefined)[]} conditions
+         */
+        const _resolveExports = (conditions) => {
+          return resolveExports(
+            exportsRootValue,
+            // @ts-expect-error ts still doesn't understand `filter(Boolean)`
+            conditions.filter(Boolean),
+            exportsPath,
+          )
+        }
 
         // NOTE: got lazy. here we check for the import/require result in different environments
         // to make sure we cover possible cases. however, a better way it to resolve the exports
         // and scan also the possible environment conditions, and return an array instead.
         for (const env of [undefined, 'node', 'browser', 'worker']) {
+          const importResult = _resolveExports(['import', env])
+          const requireResult = _resolveExports(['require', env])
+          const isDualPublish =
+            importResult &&
+            requireResult &&
+            importResult.value !== requireResult.value
+
           for (const format of ['import', 'require']) {
-            const result = resolveExports(
-              exportsRootValue,
-              // @ts-expect-error till this day, ts still doesn't understand `filter(Boolean)`
-              ['types', format, env].filter(Boolean),
-              exportsPath,
-            )
+            // the types resolved result for the corresponding js
+            const typesResult = _resolveExports(['types', format, env])
+            if (!typesResult) continue
 
-            if (!result) continue
-
-            // check if we've seen this resolve before. we also key by format as we want to distinguish
-            // incorrect exports, but only when the "exports -> path" contains that format, otherwise
-            // it's intentional fallback behaviour by libraries and we don't want to trigger a false alarm.
-            // e.g. libraries that only `"exports": "./index.mjs"` means it's ESM only, so we don't key
-            // the format, so the next run with `"require"` condition is skipped.
-            // different env can share the same key as code can usually be used for multiple environments.
+            // cache by the types path to help deduplicate the linting if we've already done so
+            // for the same environment or format. if it's dual publishing, we want to lint both times
+            // so we add the `format` to the key here.
             const seenKey =
-              result.path.join('.') + (result.dualPublish ? format : '')
+              typesResult.path.join('.') + (isDualPublish ? format : '')
             if (seenResolvedKeys.has(seenKey)) continue
             seenResolvedKeys.add(seenKey)
 
-            const resolvedPath = vfs.pathJoin(pkgDir, result.value)
             // if path doesn't exist, let the missing file error message take over instead
-            if (!(await vfs.isPathExist(resolvedPath))) continue
+            const typesResolvedPath = vfs.pathJoin(pkgDir, typesResult.value)
+            if (!(await vfs.isPathExist(typesResolvedPath))) continue
 
-            if (isDtsFile(result.value)) {
+            if (isDtsFile(typesResult.value)) {
               // if we have resolve to a dts file, it might not be ours because typescript requires
               // `.d.mts` and `.d.cts` for esm and cjs (`.js` and nearest type: module behaviour applies).
               // check if we're hitting this case :(
               const dtsActualFormat = await getDtsFilePathFormat(
-                resolvedPath,
+                typesResolvedPath,
                 vfs,
               )
 
@@ -973,21 +1009,14 @@ export async function core({ pkgDir, vfs, level, strict, _packedFiles }) {
               // only run this if not dual publish since we know dual publish should have both ESM and CJS
               // versions of the dts file, and we don't need to be lenient.
               // NOTE: could there be setups with CJS code and ESM types? seems a bit weird.
-              if (!result.dualPublish) {
-                const nonTypesResult = resolveExports(
-                  exportsRootValue,
-                  // @ts-expect-error till this day, ts still doesn't understand `filter(Boolean)`
-                  [format, env].filter(Boolean),
-                  exportsPath,
-                )
-                if (nonTypesResult?.value) {
-                  const nonTypesResolvedPath = vfs.pathJoin(
-                    pkgDir,
-                    nonTypesResult.value,
-                  )
-                  if (await vfs.isPathExist(nonTypesResolvedPath)) {
+              if (!isDualPublish) {
+                const jsResult =
+                  format === 'import' ? importResult : requireResult
+                if (jsResult) {
+                  const jsResolvedPath = vfs.pathJoin(pkgDir, jsResult.value)
+                  if (await vfs.isPathExist(jsResolvedPath)) {
                     dtsExpectFormat = await getFilePathFormat(
-                      nonTypesResolvedPath,
+                      jsResolvedPath,
                       vfs,
                     )
                   }
@@ -1003,22 +1032,22 @@ export async function core({ pkgDir, vfs, level, strict, _packedFiles }) {
 
               if (dtsActualFormat !== dtsExpectFormat) {
                 messages.push({
-                  code: 'EXPORT_TYPES_INVALID_FORMAT',
+                  code: 'EXPORTS_TYPES_INVALID_FORMAT',
                   args: {
                     condition: format,
                     actualFormat: dtsActualFormat,
                     expectFormat: dtsExpectFormat,
-                    actualExtension: vfs.getExtName(result.value),
+                    actualExtension: vfs.getExtName(typesResult.value),
                     expectExtension: getDtsCodeFormatExtension(dtsExpectFormat),
                   },
-                  path: result.path,
+                  path: typesResult.path,
                   type: 'warning',
                 })
               }
             } else {
               // adjacent dts file here is always in the correct format
               const hasAdjacentDtsFile = await vfs.isPathExist(
-                vfs.pathJoin(pkgDir, getAdjacentDtsPath(result.value)),
+                vfs.pathJoin(pkgDir, getAdjacentDtsPath(typesResult.value)),
               )
               // if there's no adjacent dts file, it's likely they don't support moduleResolution: bundler.
               // try to provide a warning.
@@ -1045,7 +1074,7 @@ export async function core({ pkgDir, vfs, level, strict, _packedFiles }) {
                       ? undefined
                       : getDtsCodeFormatExtension(dtsExpectFormat),
                   },
-                  path: result.path,
+                  path: typesResult.path,
                   type: 'warning',
                 })
               }
@@ -1075,5 +1104,65 @@ export async function core({ pkgDir, vfs, level, strict, _packedFiles }) {
       // TODO: handle nested exports key
     }
     return typesFilePath
+  }
+
+  /**
+   * @param {any} binValue
+   * @param {string[]} currentPath
+   */
+  function crawlBin(binValue, currentPath) {
+    if (typeof binValue === 'string') {
+      promiseQueue.push(async () => {
+        const binPath = vfs.pathJoin(pkgDir, binValue)
+        const binContent = await readFile(binPath, currentPath, [
+          '.js',
+          '/index.js',
+        ])
+        if (binContent === false) return
+        // Skip checks if file is not lintable
+        if (!isFilePathLintable(binValue)) return
+
+        // Check that file has shebang
+        if (!startsWithShebang(binContent)) {
+          messages.push({
+            code: 'BIN_FILE_NOT_EXECUTABLE',
+            args: {},
+            path: currentPath,
+            type: 'error',
+          })
+        }
+
+        // Check format of file
+        const actualFormat = getCodeFormat(binContent)
+        const expectFormat = await getFilePathFormat(binPath, vfs)
+        if (
+          actualFormat !== expectFormat &&
+          actualFormat !== 'unknown' &&
+          actualFormat !== 'mixed'
+        ) {
+          const actualExtension = vfs.getExtName(binPath)
+          messages.push({
+            code: isExplicitExtension(actualExtension)
+              ? 'FILE_INVALID_EXPLICIT_FORMAT'
+              : 'FILE_INVALID_FORMAT',
+            args: {
+              actualFormat,
+              expectFormat,
+              actualExtension,
+              expectExtension: getCodeFormatExtension(actualFormat),
+            },
+            path: currentPath,
+            type: 'warning',
+          })
+        }
+      })
+    } else if (typeof binValue === 'object') {
+      for (const key in binValue) {
+        const binPath = currentPath.concat(key)
+        // Nested commands are not allowed
+        if (!ensureTypeOfField(binValue[key], ['string'], binPath)) continue
+        crawlBin(binValue[key], binPath)
+      }
+    }
   }
 }